Add initial spdlog functionlity
* Setup spdlog
* Adopt config for console color
* Map loglevels
* Add some loggin macros

spdlog writes into a file and to stderr.
Filename can already be configured.
Loglevels can be used.

It is now possible to write out messages via spdlog.

Attention: since spdlog uses the same config values like the
existin log for the log file, and the existing log is not
deactivated, there might be some colision in the messages.